Add NuGet package and DLL
closes #53
