+++ .gitignore
@@ -1,13 +1,17 @@
 *.class
 *.pyc
 __pycache__/
+bin/
 blib/
 build/
 dist/
+Example.exe
 Makefile
 Minecraft-ServerStatus-*.tar.gz
+MineStat.dll
 minestat.egg-info/
 MYMETA.json
 MYMETA.yml
+obj/
 pm_to_blib
 tests/
