+++ CSharp/Example.cs
@@ -1,4 +1,5 @@
 using System;
+using MineStatLib;
 
 class Example
 {
