+++ CSharp/MineStat.sln
@@ -1,0 +1,24 @@
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30114.105
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MineStat", "MineStat\MineStat.csproj", "{11EDD760-FC1E-4BC0-8ED4-8F6D8FF97779}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{11EDD760-FC1E-4BC0-8ED4-8F6D8FF97779}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{11EDD760-FC1E-4BC0-8ED4-8F6D8FF97779}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{11EDD760-FC1E-4BC0-8ED4-8F6D8FF97779}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{11EDD760-FC1E-4BC0-8ED4-8F6D8FF97779}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {F28FBE13-E0E2-4368-87DD-3463A332C55E}
+	EndGlobalSection
+EndGlobal
